feat: add snapshot checking with conditions for controllers (#120)
* feat: add snapshot checking with conditions for controllers

We have three main controllers that create snapshots after
reconciliation is done. Since these names are predetermined they can be
used to pick up if a snapshot has already been created or not.

The next stage, uses the Ready condition state to check if a snapshot
has been reconciled or not. If not, it's possible that it has to be
recreated or pushed over because content is incorrect. In that case,
reconciliation will proceed. If the snapshot exists and it's Ready we
requeue the {Localization,Configuration,Resource}.

* adding should reconcile to all the controllers

* adding snapshot failure cases

* added ability to check if any calls have been made

* Switching to track reconciled component version instead

* Updated manifests and removed stopping the snapshot reconcile

* add version to reconciled log and add log for snapshot being done

* Added test for failed condition

* Added explanation comment on should reconcile