feat: add creating HelmRepository type sources
This change introduces providers which can handle a specific resource
type and create a specific Source object type based on the type that
each individual provider supports. The provider pattern is handy in here
to avoid having to use switch cases. Rather any given provider can
handle a specific type and if not, it will call the next provider in a
chain of providers until one is reached that can either handle the type
or any type.

If no more providers are found it will error and return that the given
type cannot be handled.