refactor: localization and mutation
This commit refactors the api for mutation so that it is cleaner; see
`examples/basic/loc.yaml` for an example of the new structure. The
mutation reconciler is also overhauled by moving the shouldReconcile
behaviour to dedicated predicates functions.

We declare a MutationObject interface with `GetSpec()` and
`GetSnapshotDigest()` methods. This allows us to watch snapshots and
also define/handle the mutation spec in a consistent way.

We now set status fields for the mutation reconcile looper beside
the relevant code.

Alongside predicates that should trigger reconciliation whenever a
dependency updates, we also add readiness checks to ensure those dependencies are
ready.

Finally we add e2e tests, the reconcile unit tests have been bypassed to
be removed or replaced in the future.

Signed-off-by: Piaras Hoban <phoban01@gmail.com>

Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>
Signed-off-by: Piaras Hoban <phoban01@gmail.com>