Add a new SnapshotWriter to MutationReconcileLooper
This commit adds a new SnapshotWriter field to the MutationReconcileLooper struct and uses it in the ReconcileMutationObject method to write a snapshot of the object after mutation. The SnapshotWriter is initialized with a new OCIWriter that writes snapshots to an OCI registry using the go-containerregistry library. The SnapshotWriter is also passed to the LocalizationReconciler and the ConfigurationReconciler in their respective tests.

Signed-off-by: Piaras Hoban <phoban01@gmail.com>