refactor: cv reconciler (#205)
* refactor: component-version Reconcile loop

This is a small refactor to help ensure the component version reconciler doesn't
enter a stalled state when the version check fails. One common scenario
where the version check may fail is if the component version has not
been pushed to a repository. The controller should handle this
gracefully and continue to check until such time as the version is
available.

Signed-off-by: Piaras Hoban <phoban01@gmail.com>

* clean up test and ready status logic

Signed-off-by: Piaras Hoban <phoban01@gmail.com>

---------

Signed-off-by: Piaras Hoban <phoban01@gmail.com>