chore: add more log output to the e2e tests (#724)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

In an effort to better debug the e2e test failures, let's dump the flux
controller logs as well after each test. The teardown should always run
hopefully because we never `FailNow()` in the e2e framework.

If this isn't working on a failure, I'll have to decise an
envtest.AfterEachTest version of the dump.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

---------

Signed-off-by: Gergely Brautigam <gergely.brautigam@sap.com>
Co-authored-by: Jakob Möller <contact@jakob-moeller.com>